pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "isr-trade-study"
version = "0.2.0"
description = "Scenario-based ISR drone deployment trade study and policy comparison harness."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Josh Lee" }]
keywords = ["isr", "operations-analysis", "trade-study", "simulation", "anduril"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "duckdb>=1.0",
]

[project.urls]
Homepage = "https://github.com/joshleh/isr-drone-deployment-trade-study"
Issues = "https://github.com/joshleh/isr-drone-deployment-trade-study/issues"

[tool.setuptools.packages.find]
where = ["src"]

